Create 'derLengthToLength()' as a module import.
